修改后的user实体字符串属性与usermapper.xml文件中的user实体映射关系一致

+++ SpringBoot2/src/main/java/com/springboot/springboot2/pojo/User.java
@@ -22,7 +22,7 @@ public class User {
     private String password;
 
     @Schema(description = "用户类型：1-管理员，2-物业，3-业主", allowableValues = {"1","2","3"}, example = "3")
-    private Integer userType;
+    private Long userType;
 
     @Schema(description = "用户ID（主键）")
     private Integer userId;
@@ -34,10 +34,10 @@ public class User {
     private Short age;
 
     @Schema(description = "用户状态：1-正常，2-禁用，3-未激活", allowableValues = {"1","2","3"}, example = "3")
-    private Integer userStatus;
+    private Long userStatus;
 
     @Schema(description = "用户所属房屋ID（下拉选择有效房间）", example = "11")
-    private Integer userRoomId;
+    private Short userRoomId;
 
     @Schema(description = "身份证号", example = "510101200001010001")
     private String idcard;
